[added] next method to observer
Imported from [this pen](http://codepen.io/appsforartists/pen/MbrmPX?editors=0010)

To be interoperable with Observables created in other libraries (and eventually, the TC39 proposal), turned the simple dispatch function into an object with a next method.  This also gives us a place to put any functionality we may need in the future.