[added] index.js to git
This repo is meant to be a learning resource as much as it is the basis for Material Motion.  Pundle strips all the whitespace in its output, so I'm adding the index.js file to git to make it easy for people researching Observables to read our source without being distracted by the type system.