[removed] observable._listeners
We originally called the creator for only the first subscription.  However, if `next` was synchronously called inside the creator, only the first observer would receive it:

```javascript
const test$ = new IndefiniteObservable(
  (observer) => {
    observer.next(Math.random());
  }
);

test$.subscribe(
  value => console.log('value 1 is', value)
);

test$.subscribe(
  value => console.log('value 2 is', value)
);
```

So, we've removed it.  This means that each subscription effectively gets its own observable (since the Observable function itself is just a wrapper around the creator).  If this turns out to be a performance issue later, we'll address it then.